refactor(7): extraer cmd_login/status/optimize/history/blacklist/help
- jobhunter/cli/login.py: wrapper + mensajes siguiente-paso
- jobhunter/cli/status.py: tabla config + stats
- jobhunter/cli/optimize.py: agente + diff visual + confirmacion
- jobhunter/cli/history.py: filtros last/company/since/all
- jobhunter/cli/blacklist.py: add/remove/list case-insensitive
- jobhunter/cli/help.py: tabla comandos + opciones + ejemplos
- job.py re-exporta todos
- tests: status (config completa/vacia), history (empty/filters/invalid
  date/last), blacklist (add/dup/remove/list)
